Seguimos con base de datos

+++ app/src/main/java/com/example/footballmatchmanager/JugadorBase.kt
@@ -4,7 +4,7 @@ open class JugadorBase(
     var valoracion: Double = 0.0,
     var nombre: String = "",
     var posicion: String = "",
-    var goles: Int? = 0, // Cambiado a Int anulable
-    var asistencias: Int? = 0 ,// Cambiado a Int anulable,
+    var goles: Long ? = 0, // Cambiado a Int anulable
+    var asistencias: Long ? = 0 ,// Cambiado a Int anulable,
 
 )
