+++ app/src/main/java/com/example/footballmatchmanager/Portero.kt
@@ -6,7 +6,7 @@ class Portero(
     valoracion: Double,
     nombre: String,
     posicion: String,
-    goles: Int = 0,
-    asistencias: Int = 0
+    goles: Long = 0,
+    asistencias: Long = 0
 ) : JugadorBase(valoracion, nombre, posicion, goles, asistencias)
